Más cambios

+++ src/bonus/scripts/install.sh
@@ -4,7 +4,7 @@ set -e
 
 # Swapfile
 if [ ! -f /swapfile ]; then
-	fallocate -l 1G /swapfile
+	fallocate -l 2G /swapfile
 	chmod 600 /swapfile
 	mkswap /swapfile
 	swapon /swapfile
@@ -58,7 +58,7 @@ if ! command -v k3d &> /dev/null; then
 	curl -s https://raw.githubusercontent.com/k3d-io/k3d/main/install.sh | bash
 fi
 
-# ArgoCD
+# ArgoCD Cli
 if ! command -v argocd &> /dev/null; then
 	curl -sSL -o argocd-linux-amd64 https://github.com/argoproj/argo-cd/releases/latest/download/argocd-linux-amd64
 	install -m 555 argocd-linux-amd64 /usr/local/bin/argocd
@@ -68,27 +68,42 @@ fi
 # K3d
 if ! k3d cluster list | grep -q "iot-cluster"; then
 	k3d cluster create iot-cluster \
-		--k3s-arg "--disable=traefik@server:0" \
 		--k3s-arg "--disable=metrics-server@server:0" \
-		--port "80:80@loadbalancer"
+		--port "80:80@loadbalancer" \
+		--port "443:443@loadbalancer"
 	mkdir -p /home/vagrant/.kube
 	cp /root/.kube/config /home/vagrant/.kube/config
 	chown vagrant:vagrant /home/vagrant/.kube/config
 fi
 
 # Namespaces
+kubectl create namespace gitlab --dry-run=client -o yaml | kubectl apply -f -
 kubectl create namespace argocd --dry-run=client -o yaml | kubectl apply -f -
 kubectl create namespace dev --dry-run=client -o yaml | kubectl apply -f -
 
+# GitLab
+if ! helm list -n gitlab | grep -q "gitlab"; then
+	helm repo add gitlab https://charts.gitlab.io/
+	helm repo update
+	helm upgrade --install gitlab gitlab/gitlab \
+		--namespace gitlab \
+		--values /tmp/k3s_config/gitlab-values.yaml \
+		--timeout 600s \
+		--wait
+fi
+
 # ArgoCD
-kubectl apply -n argocd --server-side --force-conflicts -f https://raw.githubusercontent.com/argoproj/argo-cd/stable/manifests/install.yaml
-kubectl wait --for=condition=available --timeout=300s deployment -l app.kubernetes.io/name=argocd-server -n argocd
-kubectl apply -f /tmp/k3s_config/application.yaml
-ARGOCD_PASSWORD=$(kubectl get secret argocd-initial-admin-secret -n argocd -o jsonpath="{.data.password}" | base64 -d)
-if ! ss -tlnp | grep -q ':8080'; then
+if ! kubectl get deployment -n argocd argocd-server &> /dev/null; then
+	kubectl apply -n argocd --server-side --force-conflicts -f https://raw.githubusercontent.com/argoproj/argo-cd/stable/manifests/install.yaml
+	kubectl wait --for=condition=available --timeout=300s deployment -l app.kubernetes.io/name=argocd-server -n argocd
+	kubectl patch deployment argocd-server -n argocd --type='json' -p='[{"op": "add", "path": "/spec/template/spec/containers/0/args/-", "value": "--insecure"}]'
+	kubectl apply -f /tmp/k3s_config/application.yaml
+	kubectl apply -f /tmp/k3s_config/argocd-ingress.yaml
+
+	ARGOCD_PASSWORD=$(kubectl get secret argocd-initial-admin-secret -n argocd -o jsonpath="{.data.password}" | base64 -d)
 	kubectl port-forward svc/argocd-server -n argocd 8080:443 --address 0.0.0.0 &> /dev/null &
 	sleep 5
+	argocd login localhost:8080 --username admin --password $ARGOCD_PASSWORD --insecure
+	argocd account update-password --current-password $ARGOCD_PASSWORD --new-password "1234567890"
+	kill %1
 fi
-argocd login localhost:8080 --username admin --password $ARGOCD_PASSWORD --insecure
-argocd account update-password --current-password $ARGOCD_PASSWORD --new-password "1234567890"
-kill %1
